requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pytesseract==0.3.10
Pillow==10.0.0
googletrans==4.0.0-rc1
pymongo==4.5.0
Werkzeug==2.0.1
python-dotenv==0.19.0
pdf2image==1.16.3
PyMuPDF==1.23.5
opencv-python-headless==4.8.0.74
numpy==1.24.0
gunicorn==20.1.0